pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enlumen"
version = "0.3.0"
description = "Open-source AI agent engine. Modular. No limits."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Gabo Urrutia", email = "gabogabucho@gmail.com" },
]
maintainers = [
    { name = "Gabo Urrutia", email = "gabogabucho@gmail.com" },
]
keywords = ["ai", "agent", "modular", "assistant", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "litellm>=1.40.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "jinja2>=3.1.0",
    "websockets>=12.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "build",
    "twine",
]

[project.urls]
Homepage = "https://github.com/gabogabucho/lumen-agent"
Repository = "https://github.com/gabogabucho/lumen-agent"
Issues = "https://github.com/gabogabucho/lumen-agent/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.scripts]
lumen = "lumen.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["lumen"]
artifacts = [
    "lumen/core/consciousness.yaml",
    "lumen/connectors/built-in.yaml",
    "lumen/channels/templates/**/*",
    "lumen/locales/**/*",
    "lumen/modules/**/*",
    "lumen/skills/**/*",
    "lumen/catalog/**/*",
    "logo.png",
]

[tool.hatch.build.targets.sdist]
artifacts = [
    "lumen/core/consciousness.yaml",
    "lumen/connectors/built-in.yaml",
    "lumen/channels/templates/**/*",
    "lumen/locales/**/*",
    "lumen/modules/**/*",
    "lumen/skills/**/*",
    "lumen/catalog/**/*",
    "logo.png",
]
